pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tunalab"
version = "0.1.0"
description = "Composable ML experiment infrastructure"
dependencies = [
    "torch",
    "numpy",
    "pyyaml",
    "tqdm",
]

[project.optional-dependencies]
llm = ["litellm"]
dev = ["pytest", "matplotlib"]
all = ["tunalab[llm,dev]"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests/"]